refactor: 당첨 번호 입력에 대한 static 검증 메서드 위치 변경
- 로또 상금 정보에 대한 객체 constants로 이동
