+++ src/js/utils/constants.js
@@ -4,3 +4,11 @@ export const LOTTO = Object.freeze({
   MAX_NUM: 45,
   LENGTH: 6,
 });
+
+export const LOTTO_REWARD = Object.freeze({
+  FIRST: 2000000000,
+  SECOND: 300000000,
+  THIRD: 1500000,
+  FOURTH: 50000,
+  FIFTH: 5000,
+});
